base user dto
