+++ src/main/java/tqs/sparkflow/userservice/dto/UserUpdateDTO.java
@@ -1,58 +1,5 @@
 package tqs.sparkflow.userservice.dto;
 
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-
-/**
- * Data Transfer Object for updating user information.
- */
-public class UserUpdateDTO {
-
-    @NotBlank(message = "Username is required")
-    @Size(min = 3, max = 50, message = "Username must be between 3 and 50 characters")
-    private String username;
-
-    @NotBlank(message = "Email is required")
-    @Email(message = "Invalid email format")
-    private String email;
-
-    @NotBlank(message = "Password is required")
-    @Size(min = 6, message = "Password must be at least 6 characters")
-    private String password;
-
-    private boolean operator;
-
-    // Getters and Setters
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public boolean isOperator() {
-        return operator;
-    }
-
-    public void setOperator(boolean operator) {
-        this.operator = operator;
-    }
+public class UserUpdateDTO extends BaseUserDTO {
+    // No additional fields for now; inherits everything from BaseUserDTO
 }
